requirements.txt
mysql-connector-python==8.0.22
Faker==25.3.0
tqdm==4.66.1